adjust the mysql sink
